requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
python-dotenv 
flask
mistralai
notion-Client
streamlit
streamlit-rec
python-dotenv
pydub
groq

